Cloud_Detection_GUI/requirements.txt
PyQt6
pyserial
numpy
